requirements.txt
torch>=1.12.0
numpy>=1.21.5
opencv-python>=4.5.1.48
scikit-learn>=1.0.2
scipy>=1.7.3
scikit-image>=0.19.3
Pillow>=9.2.0
torchvision>=0.13.0
matplotlib>=3.5.3
panel>=0.14.4
gudhi>=3.8.0
hausdorff>=0.2.6
requests>=2.28.1